Customize `WithRemoteData` by providing components used to render pending and error states.
Extract `DefaultErrorComponent` and `DefaultPendingComponent` which are default implementations.
