+++ src/DefaultPendingComponent.tsx
@@ -1,0 +1,3 @@
+import { FC } from 'react';
+
+export const DefaultPendingComponent: FC = () => <div>...</div>;
