+++ src/index.ts
@@ -4,3 +4,5 @@ export { RemoteData } from './RemoteData';
 export { RemoteDataStore } from './RemoteDataStore';
 export { RemoteDataStores } from './RemoteDataStores';
 export { WithRemoteData } from './WithRemoteData';
+export { DefaultErrorComponent, ErrorProps } from './DefaultErrorComponent';
+export { DefaultPendingComponent } from './DefaultPendingComponent';
